열석퍼즐 1차 제작(1쪽만 움직임)

+++ Revenant/Assets/Script/PC/FSM/Player.cs
@@ -120,8 +120,8 @@ public class Player : MonoBehaviour
 
     //스태프의 getset
     public C_STATE GetStaffState() { return staff.GetComponent<PlayerStaff>().GetState();}
-    public float GetStaffCryNumber() { return staff.GetComponent<PlayerStaff>().GetCrystalNum(); }
-    public void ChangeStaffNum(float num) { staff.GetComponent<PlayerStaff>().ChangeNum(num); }
+    public int GetStaffCryNumber() { return staff.GetComponent<PlayerStaff>().GetCrystalNum(); }
+    public void ChangeStaffNum(int num) { staff.GetComponent<PlayerStaff>().ChangeNum(num); }
     public void ChangeStaffMaterial() { staff.GetComponent<PlayerStaff>().ChangeMaterial(); }
     public void ChangeStaffMaterial(Material mat) { staff.GetComponent<PlayerStaff>().ChangeMaterial(mat); }
     public void ChangeStaffState(C_STATE state) { staff.GetComponent<PlayerStaff>().ChangeState(state); }
